refactor(converter): refactor converter (#171) (ruff + tests pass)
