+++ src/lintquarto/convert/__init__.py
@@ -1,0 +1,1 @@
+"""Conversion of QMD -> Python (-> QMD, for formatters)."""
