+++ src/lintquarto/convert/constants.py
@@ -1,0 +1,7 @@
+"""Constants used when converting."""
+
+SPACING_RULE_LINTERS = ["flake8", "ruff", "pycodestyle"]
+
+NO_LINE_COUNT_PRESERVATION = ["radon-raw"]
+
+FORMAT_SEPARATOR_PREFIX = "# %%LINTQUARTO-BLOCK-"
